refactor: rename store.rs to games_store.rs
