+++ src-tauri/src/services/mod.rs
@@ -1,4 +1,5 @@
 pub mod discord;
+pub mod games_store;
+pub mod settings_store;
 pub mod playtime;
-pub mod store;
 pub mod vndb;
